refactor: reorganize llm and trading modules
