+++ scripts/e2e_test.sh
@@ -12,16 +12,16 @@ python -m sentimental_cap_predictor.data.ingest "$TICKER" --period 1Y --interval
 python -m sentimental_cap_predictor.modeling.train_eval "$TICKER"
 
 # 3) Optimizer (optional) → writes best JSON
-python -m sentimental_cap_predictor.trader_utils.strategy_optimizer optimize "data/processed/${TICKER}_prices.csv" --iterations 200 --seed 1 || true
+python -m sentimental_cap_predictor.trading.trader_utils.strategy_optimizer optimize "data/processed/${TICKER}_prices.csv" --iterations 200 --seed 1 || true
 
 # 4) Strategy signals (optional)
-python -m sentimental_cap_predictor.strategies.moving_average gen \
+python -m sentimental_cap_predictor.trading.strategies.moving_average gen \
   --prices "data/processed/${TICKER}_prices.csv" \
   --short 10 --long 30 \
   --out "data/processed/${TICKER}_signals.csv" || true
 
 # 5) Backtest (optional)
-python -m sentimental_cap_predictor.backtest.engine run \
+python -m sentimental_cap_predictor.trading.backtest.engine run \
   --prices "data/processed/${TICKER}_prices.csv" \
   --signals "data/processed/${TICKER}_signals.csv" \
   --commission-bps 10 --slippage-bps 10 --size 1.0 || true
