+++ src/sentimental_cap_predictor/__init__.py
@@ -1,1 +1,3 @@
-from . import config  # noqa: F401
+from . import config, llm_core, trading  # noqa: F401
+
+__all__ = ["config", "llm_core", "trading"]
