+++ src/sentimental_cap_predictor/llm_core/__init__.py
@@ -1,0 +1,13 @@
+"""Core LLM components: chatbot, NLU, memory and connectors."""
+
+from . import chatbot, chatbot_frontend, chatbot_nlu, connectors, memory_indexer, llm_providers, config_llm  # noqa: F401
+
+__all__ = [
+    "chatbot",
+    "chatbot_frontend",
+    "chatbot_nlu",
+    "connectors",
+    "memory_indexer",
+    "llm_providers",
+    "config_llm",
+]
