+++ src/sentimental_cap_predictor/llm_core/connectors/__init__.py
