+++ src/sentimental_cap_predictor/llm_core/connectors/arxiv_connector.py
