+++ src/sentimental_cap_predictor/llm_core/connectors/edgar_connector.py
