+++ src/sentimental_cap_predictor/llm_core/connectors/fred_connector.py
