+++ src/sentimental_cap_predictor/llm_core/connectors/github_connector.py
