+++ src/sentimental_cap_predictor/llm_core/connectors/openalex_connector.py
