+++ src/sentimental_cap_predictor/llm_core/llm_providers/qwen_local.py
