+++ src/sentimental_cap_predictor/trading/__init__.py
@@ -1,0 +1,12 @@
+"""Trading modules including backtesting, strategies and utilities."""
+
+from . import backtest, backtest_result, backtester, strategies, strategy, trader_utils  # noqa: F401
+
+__all__ = [
+    "backtest",
+    "backtest_result",
+    "backtester",
+    "strategies",
+    "strategy",
+    "trader_utils",
+]
