+++ src/sentimental_cap_predictor/trading/backtest/engine.py
