+++ src/sentimental_cap_predictor/trading/backtest_result.py
