+++ src/sentimental_cap_predictor/trading/strategies/moving_average.py
