+++ src/sentimental_cap_predictor/trading/trader_utils/strategy_optimizer.py
