+++ tests/test_chatbot_nlu_tickers.py
@@ -1,6 +1,6 @@
 import pytest
 
-from sentimental_cap_predictor.chatbot_nlu import qwen_intent
+from sentimental_cap_predictor.llm_core.chatbot_nlu import qwen_intent
 
 
 @pytest.mark.parametrize(
